restructuring in progress including #228; tests will fail
