+++ server/opendp_apps/dataset/tests/test_depositor_serializer.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from opendp_apps.analysis.models import DepositorSetupInfo
+from opendp_apps.dataset.models import DepositorSetupInfo
 from opendp_apps.dataset.serializers import DepositorSetupInfoSerializer
 from opendp_apps.model_helpers.msg_util import msgt
