+++ server/opendp_apps/terms_of_access/models.py
@@ -63,7 +63,7 @@ class TermsOfAccessLog(TimestampedModelWithUUID):
     Records a Terms of Access agreement
     """
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.PROTECT)
-    terms_of_access = models.ForeignKey(TermsOfAccess, on_delete=models.PROTECT)
+    terms_of_access = models.ForeignKey('TermsOfAccess', on_delete=models.PROTECT)
 
     verbose_name = 'Terms of Access Log'
     verbose_name_plural = 'Terms of Access Logs'
