+++ server/opendp_project/settings/development_test.py
@@ -1,8 +1,9 @@
 from opendp_project.settings.base import *
 
-LOCAL_SETUP_DIR = os.path.join(BASE_DIR, 'test_setup', 'user_uploaded_data')
-if not os.path.isdir(LOCAL_SETUP_DIR):
-    os.makedirs(LOCAL_SETUP_DIR)
+LOCAL_SETUP_DIR = os.path.join(BASE_DIR, 'test_setup')
+USER_UPLOADED_DATA_DIR = os.path.join(LOCAL_SETUP_DIR, 'user_uploaded_data')
+if not os.path.isdir(USER_UPLOADED_DATA_DIR):
+    os.makedirs(USER_UPLOADED_DATA_DIR)
 
 DATABASES = {
     'default': {
